DotDisplayer: use a special token to separate name from index
The use of `_` as token to separate the function name from the
`g`, `v`, and optionally `u` indexes complicates the automated
demangling of function names by external scripts that consume
the output of `dumpcfg` and `oslice`. This commit uses `@` as
the name/index separator and then uses `,` to separate the
indexes from one another.